Removing redundant TestMain file
